Use a companion object for a ReviewForcast factory.
