+++ app/src/main/java/com/leapsoftware/leapforwanikani/dashboard/DashboardViewModel.kt
@@ -88,35 +88,11 @@ class DashboardViewModel(
         liveData {
             when (it) {
                 is LeapResult.Success<WKReport.Summary> -> {
-                    val reviewForecast = getForecast(it.resultData)
+                    val reviewForecast = ReviewForecast.create(it.resultData)
                     emit(reviewForecast)
                 }
             }
         }
-    }
-
-    private fun getForecast(wkSummary: WKReport.Summary): ReviewForecast {
-        val forecast = mutableListOf<HourlyForecast>()
-        val calendar = Calendar.getInstance(Locale.getDefault())
-        var totalReviewCount: Int = 0
-        for ((i, reviewsAtHour) in wkSummary.data.reviews.withIndex()) {
-            // Terrible Java Data Formatting API for the user-facing hourString
-            val sdf = SimpleDateFormat("h a", Locale.getDefault())
-            val hourOfTheDay = calendar.get(Calendar.HOUR)
-            calendar.set(Calendar.HOUR, hourOfTheDay)
-            calendar.set(Calendar.MINUTE, 0)
-            val hourString: String = sdf.format(calendar.time)
-            // [0] are the reviews available now, [1] are the reviews in an hour, etc. 24 hours provided.
-            // Skip the first [0] reviews array since we want our forecast to begin one hour from now rather than the current review status
-            // Keep track of the hourString, additionalReviews added that hour, and totalReviewCount to create an HourlyForecase; i.e. a row in the UI
-            totalReviewCount += reviewsAtHour.subject_ids.size
-            if ((i > 0) && reviewsAtHour.subject_ids.isNotEmpty()) {
-                val hourlyForecast = HourlyForecast(hourString, reviewsAtHour.subject_ids.size, totalReviewCount)
-                forecast.add(hourlyForecast)
-            }
-            calendar.add(Calendar.HOUR, 1)
-        }
-        return ReviewForecast(forecast, totalReviewCount)
     }
 
     fun refreshData() {
